==> src/Tumorigenesis/01_MakeCountMatrix.Rmd <==
---
title: "Cell Calling for tumorigenesis timeline"
author: "Karsten Bach"
date: '`r Sys.Date()`'
output:
  html_notebook:
    toc: true
    toc_depth: 2
    toc_float: true
    theme: journal
    highlight: tango
    code_folding: hide
---
***

# Call Empty Drops

I use the `emptyDrops` package to distinguish cells from backgrounds.
In the plots the below, the blue line indicates the inflection point and the black one the knee point.
Most profiles look quite nice, apart from and A2.

```{r, warning=FALSE, message=FALSE}
library(DropletUtils)
library(scran)
library(BiocParallel)
library(Matrix)
library(ggplot2)
library(dplyr)
library(cowplot)
theme_set(theme_cowplot())

smps <- list.files("../../data/Tumorigenesis/CellRangerOutput/")
smps <- grep("[0-9]$",smps,value=TRUE)
fnames <- paste0("../../data/Tumorigenesis/CellRangerOutput/",smps,"/raw")

out <- bplapply(fnames, function(fle) {
    nm <- gsub("../../data/Tumorigenesis/CellRangerOutput/|/raw","",fle)
    sce <- read10xCounts(fle, col.names=TRUE,version="3")

    bcrank <- barcodeRanks(counts(sce))

    # Only showing unique points for plotting speed.
    uniq <- !duplicated(bcrank$rank)
    fplot <- data.frame("rank"=bcrank$rank[uniq],
			"total"=bcrank$total[uniq]) 

    bcplt <- ggplot(fplot, aes(x=rank, y=total)) +
			geom_point() +
			scale_x_log10() +
			scale_y_log10() +
			geom_hline(yintercept=metadata(bcrank)$inflection,
				   lty="dashed",color="blue") +
			geom_hline(yintercept=metadata(bcrank)$knee,
				   lty="dashed") +
			ggtitle(nm) +
			xlab("Barcode Rank") +
			ylab("Total UMIs")


    set.seed(42)
    limit <- 100   
    niter <- 10000
    e.out <- emptyDrops(counts(sce), lower=limit, niters=niter)

    tbl <- table(Sig=e.out$FDR <= 0.001, Limited=e.out$Limited)

    set.seed(42)
    all.out <- emptyDrops(counts(sce), lower=limit, test.ambient=TRUE, niter=niter)

    fplot <- data.frame("X"=all.out$PValue[all.out$Total <= limit & all.out$Total > 0])

    histplt <- ggplot(fplot, aes(x=X)) +
	geom_histogram(bins=20) +
	xlab("P-Value") +
	ggtitle(nm) 

    sce <- sce[,which(e.out$FDR<=0.001)]
    sce$Sample <- nm

    out <- list("SCE"=sce,
		"BCPLT"=bcplt,
		"TBL"=tbl,
		"HISTPLT"=histplt)

    return(out)
}, BPPARAM=MulticoreParam(3))

scelist <- lapply(out, function(x) x[["SCE"]])
bcplts <- lapply(out, function(x) x[["BCPLT"]])
tbls <- lapply(out, function(x) x[["TBL"]])
histplts <- lapply(out, function(x) x[["HISTPLT"]])
rm(out)
```

## Some QC on cell calling

I here include some visualisation of the calling process as a QC step.

### Barcode Plots
```{r, warning=FALSE, message=FALSE, fig.width=13, fig.height=13}
plot_grid(plotlist=bcplts)
```

### P-Value Histograms
```{r, warning=FALSE, message=FALSE, fig.width=13, fig.height=13}
plot_grid(plotlist=histplts)
```

### P-Value limits
As the P-Values are not computed analytically you might have cells with P-Values that are 'limited' in their precision. If these are called non-significant, they might only be non-significant because of the lack of precision. This was initially the case for SampleC3, for which I adopted the number of itereations for the Monte-Carlo P-Value computation. 

- There are `r sum(unlist(lapply(tbls, function(x) x[1,2]))!=0)` samples that have cells with non-significant P-Values that are limited by the iterations. 


# Concatenate 
I then concatenated all samples into one SCE object.

```{r, warning=FALSE, message=FALSE}
addSampleID <- function(sce) {
    smp <- unique(sce$Sample)
    smp <- substring(smp,7,nchar(smp))
    bcs <- substring(colnames(sce),1,(nchar(colnames(sce))-1))
    colnames(sce) <- paste0(bcs,smp)
    return(sce)
}
scelist <- lapply(scelist,addSampleID)

m <- lapply(scelist,counts)
m <- do.call(cbind,m)
pD <- lapply(scelist,colData)
pD <- do.call(rbind,pD)
pD$barcode <- rownames(pD)
fD <- lapply(scelist,rowData)
stopifnot(all(unlist(lapply(fD[-1], function(x) identical(x,fD[[1]])))))
fD <- fD[[1]]
```

# Adding some other information to the SCE

I then added some meta data to the columns and row of the SCE.

## Column Data
```{r, warning=FALSE, message=FALSE}
# Read in meta data
meta.b1 <- read.csv("../../data/misc/SLX-15931.HV5MKBBXX.s_8.contents.csv",stringsAsFactors=FALSE)
meta.b2 <- read.csv("../../data/misc/SLX-18124.HCJKGDRXX.s_1.contents.csv",stringsAsFactors=FALSE)
meta.b3 <- read.csv("../../data/misc/SLX-19122.HKJ3FDRXX.s_1.contents.csv",stringsAsFactors=FALSE)
meta <- rbind(meta.b1,meta.b2,meta.b3)
meta$Sample.name <- gsub("-","",meta$Sample.name) 

# Add more info to phenotype Data
pD <- data.frame(pD)
pD <- mutate(pD, SeqID=substr(barcode,18,nchar(barcode))) %>%
        mutate(SampleID=plyr::mapvalues(SeqID,meta$Barcode,
				  meta$Sample.name)) %>%
	mutate(Age=plyr::mapvalues(SampleID,meta$Sample.name,
			     meta$Age)) %>%
	mutate(Chip=as.factor(substr(SampleID,nchar(SampleID),nchar(SampleID))),
	       Animal=substr(SampleID,1,nchar(SampleID)-1),
	       Condition=as.factor(substr(SampleID,1,nchar(SampleID)-1)))
    
pD$Animal <- gsub("TM|MG","",pD$Animal)
pD$Batch <- plyr::mapvalues(pD$Chip, c("1","2","3","4"),
			    c(1,1,2,3))

pD$AnimalHadTumor <- pD$Animal %in% c("WKBR61.4d","WKBR75.4b")
pD$ChlodronateExperiment <- grepl("4$",pD$SeqID)
head(pD)
```

## Feature Data

```{r, warning=FALSE, message=FALSE}
# Add more info to the feature Data
mitoGenes <- read.table("../../data/misc/MitoGenes.txt")

# Ribosmal genes
library(biomaRt)
gos <- c("GO:0003735","G0:0005840","GO:0015935","GO:0015934") # constituents of ribosome
ensembl  <- useMart("ensembl",dataset="mmusculus_gene_ensembl",host="useast.ensembl.org") # at the time of writing the ensembl.org had issues
gene.data <- getBM(attributes=c('ensembl_gene_id', 'go_id'), mart=ensembl)
gene.data <- gene.data[gene.data$go_id %in% gos,]

fD$Ribosomal <- fD$ID %in% gene.data$ensembl_gene_id

#Add ribosomal proteins that have been missed
fD$Ribosomal <- fD$Ribosomal | grepl("Rpl",fD$Symbol)

# Mitochondrial
fD$Mitochondrial <- fD$ID %in% mitoGenes$V1

# Keep for hvg
fD$KeepForHvg <- !(fD$Ribosomal | fD$Mitochondrial)

fD$uniq <- scater::uniquifyFeatureNames(fD$ID,fD$Symbol)
rownames(fD) <- fD$uniq
stopifnot(identical(fD$ID,rownames(m)))
rownames(m) <- fD$uniq
head(fD)
```

## Save
Saving all data (for genes that have a total count higher than 1 accross the dataset)
```{r, warning=FALSE, message=FALSE}
sceout <- SingleCellExperiment(assays=list(counts=m),
			       colData=pD,
			       rowData=fD)

keep <- rowSums(counts(sceout))>1
sceout <- sceout[keep,]
saveRDS(sceout,file="../../data/Tumorigenesis/Robjects/SCE.rds")
```

## Session Info
```{r}
sessionInfo()
```
