==> src/Tumorigenesis/11_CellTypeInference.Rmd <==
---
title: "Celltype Inference"
author: "Karsten Bach"
date: '`r Sys.Date()`'
output:
  html_notebook:
    toc: true
    toc_depth: 3
    toc_float: true
    theme: united
    highlight: tango
    code_folding: hide
---
***

```{r, message=FALSE,warning=FALSE}
library(scran)
library(scater)
library(plyr)
library(dplyr)
library(knitr)
library(ggplot2)
library(cowplot)
library(Matrix)
library(irlba)
library(schex)
source("../functions.R")

sce <- readRDS("../../data/Tumorigenesis/Robjects/SCE_QC_norm.rds")

umap <- read.csv("../../data/Tumorigenesis/Robjects/UMAP_corrected.csv",stringsAsFactors=FALSE)[,-1]
clst <- read.csv("../../data/Tumorigenesis/Robjects/Clusters_w6.csv",stringsAsFactors=FALSE)[,-1]
anno <- left_join(umap,clst)
sce <- sce[,umap$barcode]
pD <- data.frame(colData(sce))
pD <- left_join(pD,anno[,c("barcode","UMAP1","UMAP2","Cluster")])
colData(sce) <- DataFrame(pD)
reducedDim(sce, type="UMAP") <- pD[,c("UMAP1","UMAP2")]
```

# Major Groups

First cells where grouped into four broad categories:

- Epithelial (EpCam+)
- Immune (Ptprc+)
- Fibroblasts (Col3a1)
- Stroma (Various markers)
```{r, fig.width=9, fig.height=9, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="Condition", text_by="Cluster")
```

## Epithelial
```{r, fig.width=9, fig.height=9, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="Epcam", text_by="Cluster")
epithelial <- c(27,28,16, # Basal
		8,37,25,13,34, #Luminal
		1,20) # Tumor
epiCluster <- paste0("C",epithelial)
```

## Immune
- Most cells marked by Ptprc (Cd45)
- Neutrophils marked by S100a8
- Plasma cells Jchain
- B-Cells Cd19 
```{r, fig.width=12, fig.height=4, warning=FALSE, message=FALSE}
immune <- c(2,30,33,40,50,6, # Lymphoid
	    3,14,17,21,24,26,39,45,47,48,7) # Myeloid

plotExpression(sce, x="Cluster", features=c("Ptprc"))
plotExpression(sce, x="Cluster", features=c("S100a8"))
neutros <- c(18) # not ptprc pos but S100a8
plotExpression(sce, x="Cluster", features=c("Jchain"))
plasma <- c(31) # not ptprc but Jchain
plotExpression(sce, x="Cluster", features=c("Cd19"))
bcells <- c(9) # not ptprc but Cd19
immuneCluster <- paste0("C",c(immune,neutros,plasma,bcells))
```

# Fibroblasts

- Mostly marked by Col1a1/Col3a1
- C23 might be CAFs (pericyte derived? due to their closeness,
  check paper that we reviewed)
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="Col1a1", text_by="Cluster")
fibroblasts <- c(15,19,23,29,32,35,36,4,42,44,46,49)
fibroblastsCluster <- paste0("C",fibroblasts)
```


# Stroma
- Bit more heterogeneous not a single marker
- C22 Schwann Cell (Mbp, Mpz)
- All endothelial cells below also have Cdh5
- C11, C12,43,5 cells (Emcn)
- C41 lymphatic endothelium (Nts, Lyve1)
- C11 are pericytes (Notch3, Acta2)
```{r, fig.width=12, fig.height=4, warning=FALSE, message=FALSE}
# Stroma
# 22 is Schwann Cell (Mbp, Mpz)
# 41 lymphatic endothelium (Nts, # not really Lyve1 but Nts marked the lymphatic endothelium in prev)
# 11,12,43,5 are endothelium (Emcn)
# 7 Higd1b,Rgs5,Gm13889, Rgs4 no idea, will label as stroma
# C10 and 38 are possibly pericytes
plotExpression(sce, x="Cluster", features=c("Mbp"))
plotExpression(sce, x="Cluster", features=c("Cdh5"))
plotExpression(sce, x="Cluster", features=c("Lyve1"))
plotExpression(sce, x="Cluster", features=c("Esm1"))
plotExpression(sce, x="Cluster", features=c("Emcn"))
plotExpression(sce, x="Cluster", features=c("Notch3"))
stroma <- c(22, # Schwann
	    11,12,43,5, # Endothelial
	    41, #Lymphatic Endothelium
	    10,38) # Pericyte

stromaCluster <- paste0("C",stroma)
sce$MajorGroups <- mapvalues(as.character(sce$Cluster),
		       from=c(epiCluster, immuneCluster, fibroblastsCluster, stromaCluster),
		       to=c(rep("Epithelial",length(epiCluster)),
			    rep("Immune",length(immuneCluster)),
			    rep("Fibroblast",length(fibroblastsCluster)),
			    rep("Stroma",length(stromaCluster))))
```

This then results in the following annotation of the larger groups

```{r, fig.width=9, fig.height=9, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="MajorGroups", text_by="Cluster")
```

# Finer Annotation

- Some of the groups can be resolved in smaller sub-compartments
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
tumorCells <- c(20,1)
luminal <- c(8,37,25,13,34)
basal <- c(27,28,16)

lymphoid <- c(2,30,33,40,50,6,31,9)
myeloid <- c(3,14,17,21,24,26,39,45,47,48,7,18) 

luminal <- paste0("C",luminal)
basal <- paste0("C",basal)
myeloid <- paste0("C",myeloid)
lymphoid <- paste0("C",lymphoid)
tumorCells <- paste0("C",tumorCells)
# 
sce$Groups <- mapvalues(as.character(sce$Cluster),
			from=c(lymphoid, myeloid, basal, luminal, tumorCells, fibroblastsCluster, stromaCluster),
			to=c(rep("Lymphoid",length(lymphoid)),
			     rep("Myeloid",length(myeloid)),
			     rep("Basal",length(basal)),
			     rep("Luminal",length(luminal)),
			     rep("TumorEpithel",length(tumorCells)),
			     rep("Fibroblast",length(fibroblastsCluster)),
			     rep("Stroma",length(stromaCluster))))
```

```{r, fig.width=9, fig.height=9, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="Groups", text_by="Cluster")
ggsave("UMAP_colBy_Groups.pdf",width=10,height=10)
```

# Cell Type mapping
- Cell types were inferred based on marker gene expresssion.
- Most cell types are quite obvious.
- The fibroblasts are just numbered from Fb1-Fb10, as I am not sure how to label them
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
mark <- findMarkers(sce,sce$Cluster,pval.type="all")

celltypes <- c(1:50)
celltypes <- paste0("C",celltypes)
#pDC because of Siglec-Hhttps://www.ncbi.nlm.nih.gov/pmc/articles/PMC3977341/
names(celltypes) <- c("Tm1","Nk","MxdQ","Fb1","Ec1",
		      "Tc.Cd8.1","Mp1","Hs","Bc","Peryc",
		      "Ec2","Ec3","Avp","Dc209","Fb2",
		      "BslDbl","Tam1","Np","Fb3","Tm2",
		      "Mp2","Swc","Caf1","Dcs","Lp",
		      "Tam2","Bsl1","Bsl2","Fb4","Tc.Th2Q",
		      "Pc","Fb5","Tc.Cd8.2","Avd","Fb6",
		      "Fb7","DblQ","Peryc","Dc.Ccr7","Ilc2Q",
		      "LyEc","UkQ","Ec3","Fb8","Mst",
		      "Fb9","pDC","BasoGata2","Fb10","Tc.Cd8.3")

sce$CellTypes <- mapvalues(sce$Cluster, from=celltypes, to=names(celltypes))
sort(table(sce$CellTypes), decreasing=TRUE)
plotUMAP(sce, colour_by="Groups", text_by="CellTypes")
```

# QC Check
- As this will be the final annotation, I will one last time check that none of the clusters is funny in terms of QC criteria

```{r, fig.width=10, fig.height=3, warning=FALSE, message=FALSE}
extraQC <- read.csv("../../data/Tumorigenesis/Robjects/QC_Part2.csv")[,-1]
rownames(extraQC) <- extraQC$barcode
extraQC <- extraQC[sce$barcode,]
sce$DbltScore <- extraQC$DbltScore
sce$IsPotDamaged <- extraQC$IsPotDamaged

plotColData(sce, x="CellTypes",y="UmiSums") + scale_y_log10() + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
plotColData(sce, x="CellTypes",y="GenesDetected") + scale_y_log10() + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
plotColData(sce, x="CellTypes",y="prcntMito") + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
plotColData(sce, x="CellTypes",y="DbltScore") + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
```

## Further Doublets
### BslDbl
- Has the shape of a doublet cluster and gets high doublet scores
- Expresses Basal markers and macrophage markers
- lacks a unique marker
- Will be removed
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="BslDbl")) +
    geom_point()
```

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce <- make_hexbin(sce, nbins = 150, 
    dimension_reduction = "UMAP", use_dims=c(1,2))
gene_id <-"Krt5"
gg <- schex::plot_hexbin_gene(sce, type="logcounts", gene=gene_id, 
    action="mean", xlab="UMAP1", ylab="UMAP2", 
    title=paste0("Mean of ", gene_id))
gg + theme_void()

gene_id <-"Ptprc"
gg <- schex::plot_hexbin_gene(sce, type="logcounts", gene=gene_id, 
    action="mean", xlab="UMAP1", ylab="UMAP2", 
    title=paste0("Mean of ", gene_id))
gg + theme_void()

gene_id <-"Cxcl14"
gg <- schex::plot_hexbin_gene(sce, type="logcounts", gene=gene_id, 
    action="mean", xlab="UMAP1", ylab="UMAP2", 
    title=paste0("Mean of ", gene_id))
gg + theme_void()
```

### Fb6
- Has the shape of a doublet cluster and gets high doublet scores
- Actually marked by Apod
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Fb6")) +
    geom_point()
```

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
gene_id <-"Apod"
gg <- schex::plot_hexbin_gene(sce, type="logcounts", gene=gene_id, 
    action="mean", xlab="UMAP1", ylab="UMAP2", 
    title=paste0("Mean of ", gene_id))
gg + theme_void()
```

### Fb8
- Has the shape of a doublet cluster and gets high doublet scores
- Actually marked by Sfrp5
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Fb8")) +
    geom_point()
```

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
gene_id <-"Sfrp5"
gg <- schex::plot_hexbin_gene(sce, type="logcounts", gene=gene_id, 
    action="mean", xlab="UMAP1", ylab="UMAP2", 
    title=paste0("Mean of ", gene_id))
gg + theme_void()
```

### Fb9
- Gets high doublet scores
- Expresses Cd79a and collagens
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Fb9")) +
    geom_point()
```

```{r, fig.width=12, fig.height=4, warning=FALSE, message=FALSE}
plotExpression(sce, x="CellTypes", features=c("Cd79a"))
plotExpression(sce, x="CellTypes", features=c("Col3a1"))
```

### DblQ
- a luminal basal state
- no gene that marks it
- Very low number of cells, possibly true doublets
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="DblQ")) +
    geom_point()
```

```{r, fig.width=12, fig.height=4, warning=FALSE, message=FALSE}
plotExpression(sce, x="CellTypes", features=c("Krt5"))
plotExpression(sce, x="CellTypes", features=c("Krt8"))
```

## Difficult Clusters

### MxdQ
- Contains some obvious structure, judging from the plot it looks like it clusters with some doublets

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="MxdQ")) +
    geom_point()
```

- Sub Clustering with SNN and louvain
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="MxdQ"]
m.cor <- readRDS("../../data/Tumorigenesis/Robjects/CorrectedPCA.rds")
m.cor <- m.cor[sce.sub$barcode,]

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_louvain(igr)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up")
```

- 1 are monocytes (Plac8 and Ly6c2, according tohttps://www.cell.com/cell-systems/pdf/S2405-4712(19)30029-8.pdf) 
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Ly6c2")
plotExpression(sce.sub,x="SubCluster","Plac8")
```

- 2 Expresses Epithelial genes (-> Doublet)
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Krt8")
```

- 3 are B-Cells (Mzb1)
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["4"]]),20)
plotExpression(sce.sub,x="SubCluster","Mzb1")
```

- 4 are fibroblast doublets
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["10"]]),20)
plotExpression(sce.sub,x="SubCluster","Col1a2")
plotExpression(sce.sub,x="SubCluster","Cd14")
```

- 5 are Eosinophils (Ear2)
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Ear2")
```

- 6 and 7 similar to the Tam clusters
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Cd14")
plotExpression(sce.sub,x="SubCluster","Ms4a7")
```

- 8 are Mmp12+ Mp (Mp1)
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Mmp12")
```

- 9 are a mix of B and T cells (-> Doublet)
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Cd79a")
plotExpression(sce.sub,x="SubCluster","Cd3d")
```

- 10 have no specific marker and very low count, will be removed
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(mark.sub[["10"]])
plotColData(sce.sub,x="SubCluster",y="UmiSums")
```

- This is then all saved in the original sce set
- I also have to fix the Groups and MajorGroups annotaiton for some of these
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
mono <- sce.sub$barcode[sce.sub$SubCluster=="1"]
epidbl <- sce.sub$barcode[sce.sub$SubCluster=="2"]
bcs <- sce.sub$barcode[sce.sub$SubCluster=="3"]
fbdbl <- sce.sub$barcode[sce.sub$SubCluster=="4"]
eosin <- sce.sub$barcode[sce.sub$SubCluster=="5"]
tamex1 <- sce.sub$barcode[sce.sub$SubCluster=="6"]
tamex2 <- sce.sub$barcode[sce.sub$SubCluster=="7"]
mp1ex <- sce.sub$barcode[sce.sub$SubCluster=="8"]
btdbl <- sce.sub$barcode[sce.sub$SubCluster=="9"]
rmc <- sce.sub$barcode[sce.sub$SubCluster=="10"]
sce$CellTypes[sce$barcode %in% mono] <- "Mcs"
sce$CellTypes[sce$barcode %in% epidbl] <- "EpiDbl"
sce$CellTypes[sce$barcode %in% bcs] <- "Bcs.1"
sce$CellTypes[sce$barcode %in% fbdbl] <- "FbDbl"
sce$CellTypes[sce$barcode %in% eosin] <- "Esn"
sce$CellTypes[sce$barcode %in% tamex1] <- "Tam.1"
sce$CellTypes[sce$barcode %in% tamex2] <- "Tam.2"
sce$CellTypes[sce$barcode %in% mp1ex] <- "Mp1.1"
sce$CellTypes[sce$barcode %in% btdbl] <- "BtDbl"
sce$CellTypes[sce$barcode %in% rmc] <- "Dmgd"
sce$Groups[sce$CellTypes=="Bcs.1"] <- "Lymphoid"
```

### Tc.Cd8.1Q
- Contains some obvious structure, judging from the plot it looks like it clusters with some doublets

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Tc.Cd8.1")) +
    geom_point()
```

- Sub Clustering with SNN and louvain
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="Tc.Cd8.1"]
m.cor <- readRDS("../../data/Tumorigenesis/Robjects/CorrectedPCA.rds")
m.cor <- m.cor[sce.sub$barcode,]

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_louvain(igr)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=rowData(sce.sub)[,"KeepForHvg"])
sce.sub <- make_hexbin(sce.sub, nbins = 50, 
    dimension_reduction = "UMAP", use_dims=c(1,2))
```

- They all have TCR
- not all are Cd8a positive
- 4 highest, then 2, then 1
- 5 iis highest for Cd4
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Cd3d")
plotExpression(sce.sub,x="SubCluster","Cd8a")
plotExpression(sce.sub,x="SubCluster","Cd4")
plotUMAP(sce.sub,text_by="SubCluster",colour_by="Gzma")
```

- 4 express high level of Cd8, and parts of them high levels of Gzma and Gzmb
- most likely CTLs, could also be NK cells (Klr7a)
- best marked by Xcl1
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Xcl1")
plot_hexbin_gene(sce.sub, type="logcounts", gene="Gzma", 
    action="mean")
```

- 2 are T-Cells (Cd8)
- not marked by much tbh
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["2"]]),30)
plotExpression(sce.sub,x="SubCluster","Plek")
```

- 1 are T-Cells (Cd8)
- not marked by much tbh
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["1"]]),10)
plotExpression(sce.sub,x="SubCluster","Ctla4")
```

- 5 are Tregs
- Ctla4, Foxp3, Tnfrsf18
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["5"]]),10)
plotExpression(sce.sub,x="SubCluster","Ctla4")
plotExpression(sce.sub,x="SubCluster","Tnfrsf18")
```

- 3 are probably memory T-Cells (have Ccr7, Lef1)
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["3"]]),10)
plotExpression(sce.sub,x="SubCluster","Ccr7")
```

- 6 show signature of IFN response Irf7, Iftim3, Bcl2a1b
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["6"]]),10)
plotExpression(sce.sub,x="SubCluster","Irf7")
plotExpression(sce.sub,x="SubCluster","Ifitm3")
```

- 7 hmm maybe peripheral ? (Blk)
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["7"]]),30)
plotExpression(sce.sub,x="SubCluster","Tmem176b")
plotExpression(sce.sub,x="SubCluster","Blk")
```

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
TcCd81 <- sce.sub$barcode[sce.sub$SubCluster=="1"]
TcCd82 <- sce.sub$barcode[sce.sub$SubCluster=="2"]
TcMem <- sce.sub$barcode[sce.sub$SubCluster=="3"]
TcCTL <- sce.sub$barcode[sce.sub$SubCluster=="4"]
TcTreg <- sce.sub$barcode[sce.sub$SubCluster=="5"]
TcIfn <- sce.sub$barcode[sce.sub$SubCluster=="6"]
TcPer <- sce.sub$barcode[sce.sub$SubCluster=="7"]
sce$CellTypes[sce$barcode %in% TcCd81] <- "TcCd81"
sce$CellTypes[sce$barcode %in% TcCd82] <- "TcCd82"
sce$CellTypes[sce$barcode %in% TcMem] <- "TcMem"
sce$CellTypes[sce$barcode %in% TcCTL] <- "TcCTL"
sce$CellTypes[sce$barcode %in% TcTreg] <- "TcTreg"
sce$CellTypes[sce$barcode %in% TcIfn] <- "TcIfn"
sce$CellTypes[sce$barcode %in% TcPer] <- "TcPer"
```


- Clean Up, removing the doublets and damaged cells
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
rmClusters <- c("Dmgd","Dbls","BslDbl","EpiDbl","FbDbl","BtDbl","DblQ","Fb9")
sce <- sce[,!sce$CellTypes %in% rmClusters]
plotUMAP(sce,text_by="CellTypes")
```

- I finally run mergeCluster again as some of the clusters from MxdQ might have wrongly been split from other cells
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
m <- logcounts(sce)
rmgenes <- rownames(sce)[!rowData(sce)$KeepForHvg]
merged <- mergeCluster(m, factor(sce$CellTypes), removeGenes=rmgenes)
unique(merged$NewCluster)
```

- This suggests merging Mp1 with Mp1.1, Tam.1 with Tam1 and Mcs with DcsQ
- I am keeping the T-Cell clusters seperatley as I know that they correspond to different states of T-Cells (unsure about TcCd81 and TcCd82, but will leave it like this for now)
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce$CellTypes[sce$CellTypes %in% c("Tam.1")] <- "Tam1"
sce$CellTypes[sce$CellTypes %in% c("Mcs")] <- "DcsQ"
sce$CellTypes[sce$CellTypes %in% c("Mp1.1")] <- "Mp1"
sce$CellTypes <- factor(sce$CellTypes)
```

# Defining a color scheme
```{r, fig.width=10, fig.height=7, warning=FALSE, message=FALSE}
epithelial <- unique(sce$CellTypes[sce$MajorGroups=="Epithelial"])

library(RColorBrewer)
col.epithelial <- brewer.pal(n=length(epithelial)+1,name="Blues")[2:(length(epithelial)+1)]
names(col.epithelial) <- epithelial

myeloid <- unique(sce$CellTypes[sce$Groups=="Myeloid"])
getPalette <- colorRampPalette(brewer.pal(9, "Oranges"))
col.myeloid <- getPalette(length(myeloid)+1)[2:(length(myeloid)+1)]
names(col.myeloid) <- myeloid

lymphoid <- unique(sce$CellTypes[sce$Groups=="Lymphoid"])
getPalette <- colorRampPalette(brewer.pal(9, "Reds"))
col.lymphoid <- getPalette(length(lymphoid)+1)[2:(length(lymphoid)+1)]
names(col.lymphoid) <- lymphoid

fibroblasts <- unique(sce$CellTypes[sce$MajorGroups=="Fibroblast"])
getPalette <- colorRampPalette(brewer.pal(9, "Greens"))
col.fibroblasts <- getPalette(length(fibroblasts)+1)[2:(length(fibroblasts)+1)]
names(col.fibroblasts) <- fibroblasts

stroma <- unique(sce$CellTypes[sce$MajorGroups=="Stroma"])
col.stroma <- brewer.pal(n=length(stroma)+1,"Greys")[2:(length(stroma)+1)]
names(col.stroma) <- stroma

colors <- c(col.epithelial, col.myeloid, col.lymphoid, col.fibroblasts,
	    col.stroma)

groupColors <- c("Blue","Orange","Red","Green","Grey50")
# names(groupColors) <- c("Epithelial","Myeloid","Lymphoid","Fibroblasts","Stroma")
# sce$GroupColors <- mapvalues(sce$Groups, names(groupColors), groupColors)
sce$Colors <- mapvalues(sce$CellTypes, names(colors),colors)

pD <- data.frame(colData(sce))
ggplot(pD, aes(x=UMAP1, y=UMAP2, color=CellTypes)) +
    geom_point() +
    scale_color_manual(values=levels(pD$Colors))
```
# Final Annotation and Save
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
saveRDS(sce,file="../../data/Tumorigenesis/Robjects/SCE_final.rds")
sessionInfo()
```
